Tratando Erro 404

+++ forum/src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -64,29 +65,38 @@ public class TopicosController {
 		
 	// DETALHANDO OS DADOS DO TOPICO
 	@GetMapping("/{id}")
-	public DetalhesDoTopicoDTO detalhar(@PathVariable Long id) {
-		Topico topico = topicoRepository.getOne(id);
-		return new DetalhesDoTopicoDTO(topico);
-	}
+	public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable Long id) {
+		Optional<Topico> topico = topicoRepository.findById(id);
+		if(topico.isPresent()) {
+		return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+		}
+		return ResponseEntity.notFound().build();
+		}
 
 	
 	// ATUALIZANDO OS DADOS DO TOPICO
 		@PutMapping("/{id}")	
 		@Transactional        // -> uTILIZADO O @TRANSACTIONAL PARA O SPRING EFETUAR O UPDATE NO BANCO
 		public ResponseEntity<TopicoDTO>  atualizar( @PathVariable Long id, @RequestBody @Valid  AtualizacaoTopicoForm  form) {
-			Topico topico = form.atualizar(id, topicoRepository);
-			
-			return ResponseEntity.ok(new TopicoDTO(topico));
-		}
+			Optional<Topico> optional = topicoRepository.findById(id);
+			if(optional.isPresent()) {
+				Topico topico = form.atualizar(id, topicoRepository);
+				return ResponseEntity.ok(new TopicoDTO(topico));
+			}
+				return ResponseEntity.notFound().build();
+			}
+		
 	
 	
 		// DELETANDO OS DADOS DO TOPICO	
 		@DeleteMapping("/{id}")
 		@Transactional
 		public ResponseEntity<?> remover(@PathVariable Long id){
-			topicoRepository.deleteById(id);
-			return ResponseEntity.ok().build();
+			Optional<Topico> optional = topicoRepository.findById(id);
+				if(optional.isPresent()) {
+					topicoRepository.deleteById(id);
+					return ResponseEntity.ok().build();
+				}
+				return ResponseEntity.notFound().build();
+				}
 		}
-		
-	
-}
